Client, Request: allow per-request middleware
I don't exactly fancy this patch, but here it is anyways.

This adds an optional middleware stack to `Request`, allowing
`RequestBuilder` and one-off requests to have middleware added
onto them, but also allows one-off requests with middleware to
have the middleware stack be appended to client middleware
for just that request.